Functions

rel_path_t

TuringModels.rel_path_t — Method.

relpatht

Relative path using the TuringModels src/ directory.

Example to get access to the data subdirectory

rel_path_t("..", "data")

source

generate_t

TuringModels.generate_t — Method.

generate_t

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_t(sd = script_dict_t)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

source

TuringModels.generate_t — Method.

generate_t

Generate notebooks and scripts in a single chapter.

Method

generate_t(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate("04")

source

TuringModels.generate_t — Method.

generate_t

Generate a single notebook and script

Method

generate_t(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generatet("04", "m4.1d.jl") or generatet("04/m4.1t.jl")

source

ScriptEntry

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

scriptentry

StatisticalRethinking.scriptentry — Method.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

maximum_a_posteriori

TuringModels.maximum_a_posteriori — Method.

maximum_a_posteriori

Compute the maximum_a_posteriori of a model.

Method

maximum_a_posteriori(model, lower_bound, upper_bound)

Required arguments

* `model::Turing model`
* `lower_bound::Float64`
* `upper_bound::Float64`

Return values

* `result`                       : Maximum_a_posterior vector

Examples

See m2.1t.jl

source

link to m2.1t.jl

link to maximum_a_posteriori(model, lower_bound, upper_bound)