Functions

rel_path_t

TuringModels.rel_path_t — Method.

relpatht

Relative path using the TuringModels src/ directory.

Example to get access to the data subdirectory

rel_path_t("..", "data")

source

generate_t

generate_t_(; sd=script_dict_t)
generate_t(chapter::AbstractString; sd=script_dict_t_)
generate_t(chapter::AbstractString, scriptfile::AbstractString; sd=script_dict_t_)

ScriptEntry

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

scriptentry

StatisticalRethinking.scriptentry — Method.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

maximum_a_posteriori

TuringModels.maximum_a_posteriori — Method.

maximumaposteriori

Compute the maximumaposteriori of a model.

Method

maximum_a_posteriori(model, lower_bound, upper_bound)

Required arguments

* `model::Turing model`
* `lower_bound::Float64`
* `upper_bound::Float64`

Return values

* `result`                       : Maximum_a_posterior vector

Examples

See 02/m2.1t.jl

source